rl_ws/pyproject.toml
[project]
name = "bsl-droid-rl"
version = "0.1.0"
description = "Reinforcement Learning workspace for BSL-Droid locomotion policy training"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    # Genesis physics simulation
    "genesis-world>=0.3.0",
    # PyTorch (CPU for macOS default, CUDA version should be installed separately for GPU training)
    "torch>=2.0.0",
    # RL training
    "rsl-rl-lib==2.2.4",
    "tensorboard>=2.14.0",
    # Utilities
    "numpy>=1.24.0",
    "gymnasium>=0.29.0",
    "mujoco>=3.4.0",
    "imageio>=2.37.2",
    "imageio-ffmpeg>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["envs"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
